I think i need some sleep

+++ src/controllers/css.ts
@@ -1,6 +1,6 @@
 import db from '../db';
 import {getUserByID} from './user';
-import {RunResult} from "sqlite3";
+import {isOfType, SqlRowTypeError} from "../util";
 
 export type CSS = {
     id: number,
@@ -12,27 +12,54 @@ export type CSS = {
     category: CSSCategory,
 }
 
+const isCSS = (o: unknown): o is CSS => isOfType(o, {
+    id: x => typeof x === 'number',
+    name: x => typeof x === 'string',
+    viewed_time: x => typeof x === 'number',
+    author_id: x => typeof x === 'number',
+    html: x => typeof x === 'string',
+    css: x => typeof x === 'string',
+    category: x => (cssCategories as unknown as unknown[]).includes(x),
+});
 
-export function getCSSs(ids: number[]): Promise<CSS[]> {
-    return new Promise((resolve, reject) => {
-        db.all(`SELECT * FROM css WHERE id IN (${ids.map(() => '?').join(',')})`, ids, (err: Error|null, objects: CSS[]) => {
+const pickCSS = ({id, name, viewed_time, author_id, html, css, category}: CSS): CSS => ({id, name, viewed_time, author_id, html, css, category});
+
+
+export function getCSSs(ids: number[]): Promise<CSS[] | Error> {
+    return new Promise((resolve) => {
+        db.all(`SELECT * FROM css WHERE id IN (${ids.map(() => '?').join(',')})`, ids, (err, rows) => {
             if (err !== null) {
-                reject(err);
+                console.error(err);
+                resolve(err);
             } else {
-                resolve(objects);
+                if (rows.length === 0) {
+                    resolve([]);
+                } else {
+                    if (isCSS(rows[0])) {
+                        resolve((rows as CSS[]).map(x => pickCSS(x)));
+                    } else {
+                        let err = SqlRowTypeError();
+                        console.error(err);
+                        resolve(err);
+                    }
+                }
             }
         });
     });
 }
 
 
-export function getValidIDs(options: {
+const isIDRow = (o: unknown): o is {id: number} => isOfType(o, {
+    id: x => typeof x === 'number',
+});
+
+export async function getValidIDs(options: {
     category?: string,
     author_id?: number,
     limit?: number,
     offset?: number,
     order?: string[],
-}): Promise<number[]> {
+}): Promise<number[] | Error> {
     let where = '';
     let order = '';
     let limit = '';
@@ -48,12 +75,13 @@ export function getValidIDs(options: {
             params.push(options.author_id);
         }
         where = `WHERE ${conditions.join(' AND ')} `;
-        console.log(where)
     }
     if (options.order !== undefined) {
         for (let each of options.order) {
             if (each.match(/^[a-zA-Z_]+$/) === null) {
-                throw Error('Illegal column name contained inside options.order');
+                let err = Error('Illegal column name contained inside options.order');
+                console.error(err);
+                return err;
             }
         }
         order = `ORDER BY ${options.order.join(',')} `;
@@ -66,12 +94,23 @@ export function getValidIDs(options: {
             params.push(options.offset);
         }
     }
-    return new Promise((resolve, reject) => {
-        db.all(`SELECT id FROM css ${where} ${order} ${limit}`, params, (err: Error|null, objects: CSS[]) => {
+    return new Promise((resolve) => {
+        db.all(`SELECT id FROM css ${where} ${order} ${limit}`, params, (err, rows) => {
             if (err !== null) {
-                reject(err);
+                console.error(err);
+                resolve(err);
             } else {
-                resolve(objects.map(e => e.id));
+                if (rows.length === 0) {
+                    resolve([]);
+                } else {
+                    if (isIDRow(rows[0])) {
+                        resolve((rows as {id: number}[]).map(x => x.id));
+                    } else {
+                        let err = SqlRowTypeError();
+                        console.error(err);
+                        resolve(err);
+                    }
+                }
             }
         });
     });
@@ -89,17 +128,23 @@ export const cssCategories = [
     "special effect",
 ] as const;
 export type CSSCategory = typeof cssCategories[number];
-export async function createCSS(userID: number, password_hashed: string, name: string, html: string, css: string, category: CSSCategory): Promise<number> {
+export async function createCSS(userID: number, password_hashed: string, name: string, html: string, css: string, category: CSSCategory): Promise<number | Error> {
     let user = await getUserByID(userID);
+    if (user instanceof Error) {
+        return user;
+    }
     if (user.password_hashed !== password_hashed) {
-        throw Error("Incorrect password");
+        let err = Error("Incorrect password");
+        console.error(err);
+        return err;
     }
 
-    return new Promise((resolve, reject) => {
+    return new Promise((resolve) => {
         db.run(`INSERT INTO css (name, author_id, html, css, category) VALUES (?, ?, ?, ?, ?)`, [name, userID, html, css, category],
-            function (this:RunResult, err: Error|null) {
+            function (err) {
                 if (err !== null) {
-                    reject(err);
+                    console.error(err);
+                    resolve(err);
                 } else {
                     resolve(this.lastID);
                 }
@@ -108,21 +153,33 @@ export async function createCSS(userID: number, password_hashed: string, name: s
 }
 
 
-export async function deleteCSS(id: number, password_hashed: string): Promise<void> {
-    let [style] = await getCSSs([id]);
+export async function deleteCSS(id: number, password_hashed: string): Promise<void| Error> {
+    let styles = await getCSSs([id]);
+    if (styles instanceof Error) {
+        return styles
+    }
+    let style = styles[0];
     if (style === undefined) {
-        throw Error('ID does not exist');
+        let err = Error('ID does not exist');
+        console.error(err);
+        return err;
     }
     let user = await getUserByID(style.author_id);
-
-    if (user.password_hashed !== password_hashed) {
-        throw Error("Incorrect password!");
+    if (user instanceof Error) {
+        return user;
     }
 
-    return new Promise((resolve, reject) => {
+    if (user.password_hashed !== password_hashed) {
+        let err = Error("Incorrect password!");
+        console.error(err);
+        return err;
+    }
+
+    return new Promise((resolve) => {
         db.run(`DELETE FROM css WHERE id=${id}`, [], (err) => {
             if (err !== null) {
-                reject(err);
+                console.error(err);
+                resolve(err);
             } else {
                 resolve();
             }
